pyproject.toml
[tool.poetry]
name = "dizzy"
version = "0.1.0"
description = ""
authors = ["Grayson Miller <grayson.miller124@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^6.0"
zmq = "^0.0.0"
click = "^8.1.3"
pytest = "^7.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
dizzy = "dizzy.__main__:main"
dizae = "dizzy.daemon.__main__:main"
dizgen = "dizzy.daemon.gen:main"
dizzy-server = "dizzy.daemon.__main__:server"
dizzy-client = "dizzy.daemon.__main__:client"

[build]
include = ["data/**"]